requirements.txt
Flask==1.1.2
Flask-Bootstrap4==4.0.2